requirements.txt
pandas
matplotlib
scikit-learn
rdkit-pypi
